Last Final Project
I did the best I could with this website, I used w3school for help on certain things or on how to do things but for the things I wanted to do wasn't working for the website and kept getting stuck. I really do hope I pass and graduate to make my family proud! Thank you for this wonderful opportunity! you were an awesome instructor!